+++ srcs/map/valid.c
@@ -6,18 +6,18 @@
 /*   By: kkamata <kkamata@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/06 19:31:56 by kkamata           #+#    #+#             */
-/*   Updated: 2021/10/08 11:47:18 by kkamata          ###   ########.fr       */
+/*   Updated: 2021/10/09 05:52:41 by kkamata          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../includes/so_long.h"
 
-t_bool	is_surrounded_by_wall(t_game *game, int x, int y)
+t_bool	is_surrounded_by_wall(t_game *game, int64_t x, int64_t y)
 {
 	if (x < 0 || y < 0)
 		return (FALSE);
-	if (game->map.matrix[COL] <= (size_t)x || \
-		game->map.matrix[ROW] <= (size_t)y)
+	if (game->map.matrix[COL] <= x || \
+		game->map.matrix[ROW] <= y)
 		return (FALSE);
 	if (game->map.submap[x + game->map.matrix[COL] * y] == MAP_WALL)
 		return (TRUE);
@@ -35,9 +35,9 @@ t_bool	is_surrounded_by_wall(t_game *game, int x, int y)
 
 t_bool	is_valid_elements(t_game *game)
 {
-	size_t	i;
-	size_t	exit_count;
-	size_t	player_count;
+	int64_t		i;
+	uint64_t	exit_count;
+	uint64_t	player_count;
 
 	i = 0;
 	exit_count = 0;
