+++ srcs/position/position.c
@@ -6,13 +6,13 @@
 /*   By: kkamata <kkamata@student.42tokyo.jp>       +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/07 18:50:39 by kkamata           #+#    #+#             */
-/*   Updated: 2021/10/08 11:41:18 by kkamata          ###   ########.fr       */
+/*   Updated: 2021/10/08 17:32:33 by kkamata          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../includes/so_long.h"
 
-size_t	locate_position(t_game *game, size_t position[2])
+size_t	locate_position(t_game *game, int64_t position[2])
 {
 	return (position[X] + game->map.matrix[COL] * position[Y]);
 }
